+++ src/main/java/net/lukeiscoding/minecraft/forge/opitemsmod/items/food/ModFoodProperties.java
@@ -1,0 +1,27 @@
+package net.lukeiscoding.minecraft.forge.opitemsmod.items.food;
+/*
+        Copyright (C) 2022  Luke Is Coding
+
+        This program is free software: you can redistribute it and/or modify
+        it under the terms of the GNU General Public License as published by
+        the Free Software Foundation, either version 3 of the License, or
+        (at your option) any later version.
+
+        This program is distributed in the hope that it will be useful,
+        but WITHOUT ANY WARRANTY; without even the implied warranty of
+        MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+        GNU General Public License for more details.
+
+        You should have received a copy of the GNU General Public License
+        along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.food.FoodProperties;
+
+public class ModFoodProperties {
+
+
+    public static final FoodProperties WEED_BROWNIE = (new FoodProperties.Builder()).fast().nutrition(4).saturationMod(0.1f).effect(new MobEffectInstance(MobEffects.MOVEMENT_SPEED, 1200), 12.5f).build();
+}