+++ src/main/java/net/lukeiscoding/minecraft/forge/opitemsmod/world/features/ModConfigureFeatures.java
@@ -17,14 +17,11 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 import net.lukeiscoding.minecraft.forge.opitemsmod.registry.RegisterBlocks;
-import net.lukeiscoding.minecraft.forge.opitemsmod.world.OreGeneration;
 import net.minecraft.core.Holder;
 import net.minecraft.data.worldgen.features.FeatureUtils;
 import net.minecraft.data.worldgen.features.OreFeatures;
 import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
 import net.minecraft.world.level.levelgen.feature.Feature;
-import net.minecraft.world.level.levelgen.feature.OreFeature;
-import net.minecraft.world.level.levelgen.feature.configurations.NoneFeatureConfiguration;
 import net.minecraft.world.level.levelgen.feature.configurations.OreConfiguration;
 
 import java.util.List;
@@ -33,17 +30,29 @@ public class ModConfigureFeatures {
 
     public static final List<OreConfiguration.TargetBlockState> OVER_WORLD_TOPAZ_ORE = List.of(
             OreConfiguration.target(OreFeatures.STONE_ORE_REPLACEABLES, RegisterBlocks.TOPAZ_ORE.get().defaultBlockState()),
-            OreConfiguration.target(OreFeatures.STONE_ORE_REPLACEABLES, RegisterBlocks.TOPAZ_DEEPSLATE_ORE.get().defaultBlockState())
+            OreConfiguration.target(OreFeatures.DEEPSLATE_ORE_REPLACEABLES, RegisterBlocks.TOPAZ_DEEPSLATE_ORE.get().defaultBlockState())
     );
 
     public static final List<OreConfiguration.TargetBlockState> OVER_WORLD_CARBONADO_ORE = List.of(
             OreConfiguration.target(OreFeatures.STONE_ORE_REPLACEABLES, RegisterBlocks.CARBONADO_ORE.get().defaultBlockState()),
-            OreConfiguration.target(OreFeatures.STONE_ORE_REPLACEABLES, RegisterBlocks.CARBONADO_DEEPSLATE_ORE.get().defaultBlockState())
+            OreConfiguration.target(OreFeatures.DEEPSLATE_ORE_REPLACEABLES, RegisterBlocks.CARBONADO_DEEPSLATE_ORE.get().defaultBlockState())
     );
 
-    public static final Holder<ConfiguredFeature<OreConfiguration, ?>> TOPAZ_ORE = FeatureUtils.register("topaz_ore", Feature.ORE, new OreConfiguration(OVER_WORLD_TOPAZ_ORE, 20));
-    public static final Holder<ConfiguredFeature<OreConfiguration, ?>> TOPAZ_DEEPSLATE_ORE = FeatureUtils.register("topaz_deepslate_ore", Feature.ORE, new OreConfiguration(OVER_WORLD_TOPAZ_ORE, 20));
+    public static final List<OreConfiguration.TargetBlockState> OVER_WORLD_RUBY_ORE = List.of(
+            OreConfiguration.target(OreFeatures.STONE_ORE_REPLACEABLES, RegisterBlocks.RUBY_ORE.get().defaultBlockState()),
+            OreConfiguration.target(OreFeatures.DEEPSLATE_ORE_REPLACEABLES, RegisterBlocks.RUBY_DEEPSLATE_ORE.get().defaultBlockState())
+    );
 
-    public static final Holder<ConfiguredFeature<OreConfiguration, ?>> CARBONADO_ORE = FeatureUtils.register("carbonado_ore", Feature.ORE, new OreConfiguration(OVER_WORLD_CARBONADO_ORE, 12));
+    /**
+     * @var p_161014_ = vain size
+     */
+
+    public static final Holder<ConfiguredFeature<OreConfiguration, ?>> TOPAZ_ORE = FeatureUtils.register("topaz_ore", Feature.ORE, new OreConfiguration(OVER_WORLD_TOPAZ_ORE, 16));
+    public static final Holder<ConfiguredFeature<OreConfiguration, ?>> TOPAZ_DEEPSLATE_ORE = FeatureUtils.register("topaz_deepslate_ore", Feature.ORE, new OreConfiguration(OVER_WORLD_TOPAZ_ORE, 16));
+
+    public static final Holder<ConfiguredFeature<OreConfiguration, ?>> CARBONADO_ORE = FeatureUtils.register("carbonado_ore", Feature.ORE, new OreConfiguration(OVER_WORLD_CARBONADO_ORE, 20));
     public static final Holder<ConfiguredFeature<OreConfiguration, ?>> CARBONADO_DEEPSLATE_ORE = FeatureUtils.register("carbonado_deepslate_ore", Feature.ORE, new OreConfiguration(OVER_WORLD_CARBONADO_ORE, 20));
+
+    public static final Holder<ConfiguredFeature<OreConfiguration, ?>> RUBY_ORE = FeatureUtils.register("ruby_ore", Feature.ORE, new OreConfiguration(OVER_WORLD_RUBY_ORE, 40));
+    public static final Holder<ConfiguredFeature<OreConfiguration, ?>> RUBY_DEEPSLATE_ORE = FeatureUtils.register("ruby_deepslate_ore", Feature.ORE, new OreConfiguration(OVER_WORLD_RUBY_ORE, 40));
 }