+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 ***
 
+## Version 0.0.2-alpha
+* Fixed the typo in the item namespace for Ruby Sword.
+* Fixed ores so that more of them spawn.
+* Fixed the block state for carbonado deepslate ore
+* Fixed crafting recipes for ruby sword, topaz how and shovel.
+* Changed the harvest level for ruby blocks and ores from diamond to iron.
+* Added ruby door, trapdoor, button, and pressure plate.
+* Added a portal frame block for a dimension in a future update.
+* Added a creative mode tab for redstone.
+* Added a creative mode tab for food.
+* Added a weed brownie.
+
 ## Version 0.0.1-alpha
 * The first alpha release of the mod.